requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
pyyaml>=6.0
certifi>=2023.7.22 